Crate: share node-by-path lookup; fold svg blur-group helper
`get_node_by_path` was written three times — private in takumi-core, exported
mutably from takumi-raster, and again as `node_at` in takumi-svg. Make the core
pair (`get_node_by_path` + `get_node_mut_by_path`) the one source both backends
use. In the svg backend, fold the duplicated "wrap in a blur group when the
radius is positive" block from the outset and inset shadow emitters into
`emit_with_blur`.
